OMS7-N / Substratum support: updated to 2017-01-13
Substratum expose hardcoded colors

Themes: Separate background color from text color

Signed-off-by: Bryan Owens <djbryan3540@gmail.com>

themes: InCallUI dialpad digits color

11/2016: Port to Nougat's Dialer (from marshmallow's InCallUI) by KreAch3R

InCallUI's dialpad digits (not the digits' text) color is linked to the InCallUI theme primary color, on runtime. This is coded as such because Google wanted to update the InCallUI's theme (primary and primary-dark) colors in MSIM devices on the fly, based on which SIM card was used to make the call.
This commit:
* makes this behavior configurable by a boolean
* links the digits color to a color value

This way, the default behavior is preserved. The themer has to set the "config_dialpadDigitsStaticColor" boolean to true, in order to theme the color value. If he doesn't change the boolean, the default behavior is preserved and the digits color works as Google wanted.

YAOSP-NG : place config & color xml's in InCallUI/res/values  instead of res/values

Signed-off-by: George G <kreach3r@users.noreply.github.com>

Themes: Separate background color from text color

Signed-off-by: Alex Cruz <mazdarider23@gmail.com>